seeders for listings

+++ database/migrations/2022_03_15_193602_create_listings_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string("coordinates");
             $table->text("description");
             $table->integer("size");
-            $table->integer("daily_rate");
+            $table->float("daily_rate", 8, 2);
             $table->integer("rating");
             $table->timestamps();
         });
